requirement.txt
asgiref==3.2.10
dj-database-url==0.5.0
Django==3.0
django-simple-search==1.0.2
gunicorn==20.0.4
Pillow==7.2.0
psycopg2==2.8.5
pytz==2020.1
sqlparse==0.3.1
